Float second bug fix

+++ utils.py
@@ -78,7 +78,7 @@ def record_user(silence_seconds, vad):
     while True:
         data = stream.read(CHUNK)
         frames.append(data)
-        contains_speech = vad.contains_speech(frames[-one_second_iters * silence_seconds:])
+        contains_speech = vad.contains_speech(frames[int(-one_second_iters * silence_seconds):])
         if not started and contains_speech:
             started = True
             print("*listening to speech*")
